refactor: move UIntPtr64<T> to new file
